extracted APIClient Error
